02-legal-ai-assistant/requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.3.0
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
pypdf==5.1.0
python-docx==1.1.2
python-dotenv==1.0.1
huggingface-hub==0.27.1
rich==13.9.4
